feat: show M17 metadata when it's received
Add support for showing metadata when received from the M17 protocol.
This also adds an implementation for scrolling when the text is too wide to fit on the screen.
On most devices, this will show up on the same line as where the reflector info is shown, but the protocol itself means that it will alternate just on is own.

Co-authored-by: Rick <crs026@yahoo.com>