localization: extract initial strings so that translations can begin
This change creates a dummy gettext method that in the future will be replaced by its real implementation. The goal is to identify strings here early so that translations of those strings can begin in earnest.

This change also saw a few strings change to const to be compatible with gettext.

Finally, this change introduced initial versions of the extracted PO template and initial language PO files. There is no automation or doc for this yet, but this will be added soon in the next steps.

The POT file was generated using:
```
 git ls-files | grep -E "\.(h|hpp|c|cpp)$" |xargs -r xgettext -c -o openrtx/src/locale/OpenRTX.pot --omit-header
```

Signed-off-by: Ryan Turner <ryan@turnrye.com>