protocols: M17: add metadata class
Add classes oriented around handling metadata in the M17 protocol. This
allows for a cleaner separation between the OpMode class, and simplifies
future adoption of new M17 v3 which we anticipate will significantly
change the shape of the LSF and the metadata.

This change adds a base metadata class, subclasses for the current types
of metadata that are supported by OpenRTX, and a factory for easily
creating these based on the protocol specifications.

Signed-off-by: Ryan Turner <ryan@turnrye.com>